+++ packages/utilities/psammead-styles/index.test.jsx
@@ -40,6 +40,8 @@ const fontsExpectedExports = {
   F_PADAUK_BOLD: 'function',
   F_SHONAR_BANGLA_REGULAR: 'function',
   F_SHONAR_BANGLA_BOLD: 'function',
+  F_NOTO_SERIF_BENGALI_REGULAR: 'function',
+  F_NOTO_SERIF_BENGALI_BOLD: 'function',
 };
 
 const detectionExpectedExports = {