+++ packages/utilities/psammead-styles/src/fonts.js
@@ -20,6 +20,10 @@ const baseUrlNotoSansEthiopic =
 const baseUrlPadauk =
   'https://ws-downloads.files.bbci.co.uk/fonts/Padauk/v2.8/';
 
+// update this link when font server updated
+const baseUrlNotoSerifBengali =
+  'https://ws-downloads.test.files.bbci.co.uk/fonts/NotoSerifBengali/v1.00/';
+
 const baseUrlShonarBangla =
   'https://ws-downloads.files.bbci.co.uk/fonts/ShonarBangla/v5.91/';
 
@@ -539,6 +543,38 @@ export const F_SHONAR_BANGLA_BOLD = baseUrlOverride => `
   baseUrlOverride || baseUrlShonarBangla
 }bold.eot') format('eot'), url('${
   baseUrlOverride || baseUrlShonarBangla
+}bold.ttf') format('ttf');
+    font-display: optional;
+  }
+`;
+
+// Bengali
+export const F_NOTO_SERIF_BENGALI_REGULAR = baseUrlOverride => `
+  @font-face {
+    font-family: "Noto Serif Bengali";
+    font-weight: 400;
+    font-style: normal;
+    src: url('${
+      baseUrlOverride || baseUrlNotoSerifBengali
+    }normal.woff') format('woff'), url('${
+  baseUrlOverride || baseUrlNotoSerifBengali
+}normal.eot') format('eot'),url('${
+  baseUrlOverride || baseUrlNotoSerifBengali
+}normal.ttf') format('ttf');
+    font-display: optional;
+  }`;
+
+export const F_NOTO_SERIF_BENGALI_BOLD = baseUrlOverride => `
+  @font-face {
+    font-family: "Noto Serif Bengali";
+    font-weight: 700;
+    font-style: normal;
+    src: url('${
+      baseUrlOverride || baseUrlNotoSerifBengali
+    }bold.woff') format('woff'),url('${
+  baseUrlOverride || baseUrlNotoSerifBengali
+}normal.eot') format('eot'), url('${
+  baseUrlOverride || baseUrlNotoSerifBengali
 }bold.ttf') format('ttf');
     font-display: optional;
   }