+++ packages/components/psammead-bulleted-list/CHANGELOG.md
@@ -3,6 +3,7 @@
 | Version       | Description                                                                                                                 |
 | ------------- | --------------------------------------------------------------------------------------------------------------------------- |
 | 3.1.4 | [PR#4548](https://github.com/bbc/psammead/pull/4548) Bumps dependencies |
+| 3.1.3 | [PR#4547](https://github.com/bbc/psammead/pull/4547) Bumps packages |
 | 3.1.2         | [PR#4545](https://github.com/bbc/psammead/pull/4545) Bump dependencies for @bbc/psammead-styles                             |
 | 3.1.1         | [PR#4552](https://github.com/bbc/psammead/pull/4552) Bump psammead-styles |
 | 3.1.0         | [PR#4529](https://github.com/bbc/psammead/pull/4529) Allow bullet point shape and colour to be customised                   |