+++ packages/components/psammead-embed-error/CHANGELOG.md
@@ -3,6 +3,7 @@
 | Version        | Description                                                                                                                 |
 | -------------- | --------------------------------------------------------------------------------------------------------------------------- |
 | 3.0.26 | [PR#4548](https://github.com/bbc/psammead/pull/4548) Bumps dependencies |
+| 3.0.25 | [PR#4547](https://github.com/bbc/psammead/pull/4547) Bumps packages |
 | 3.0.24 | [PR#4545](https://github.com/bbc/psammead/pull/4545) Bump dependencies for @bbc/psammead-styles                                     |
 | 3.0.23 | [PR#4552](https://github.com/bbc/psammead/pull/4552) Bump psammead-styles |
 | 3.0.22 | [PR#4497](https://github.com/bbc/psammead/pull/4497) Bump psammead-styles |